+++ pkgs/firehose/bin/health.dart
@@ -56,6 +56,16 @@ void main(List<String> arguments) async {
       help: 'The name of the file to write the resulting comment to.',
     )
     ..addFlag('local', help: 'Run locally', defaultsTo: true);
+      'license',
+      help: 'The license string to insert if missing.'
+          ' %YEAR% will be replaced with the current year',
+    )
+    ..addOption(
+      'license_test_string',
+      help:
+          'A file containing this string will be considered having a license.',
+    );
+
   for (var check in Check.values) {
     argParser.addMultiOption(
       'ignore_${check.name}',
@@ -69,16 +79,21 @@ void main(List<String> arguments) async {
       (checkStr) => Check.values.firstWhere((c) => c.displayName == checkStr));
   final warnOn = parsedArgs.multiOption('warn_on');
   final failOn = parsedArgs.multiOption('fail_on');
-  final flutterPackages = _listNonEmpty(parsedArgs, 'flutter_packages');
-  final ignorePackages = _listNonEmpty(parsedArgs, 'ignore_packages');
+  final flutterPackages = parsedArgs.listNonEmpty('flutter_packages');
+  final ignorePackages = parsedArgs.listNonEmpty('ignore_packages');
   final ignoredFor = Map.fromEntries(Check.values
-      .map((c) => MapEntry(c, _listNonEmpty(parsedArgs, 'ignore_${c.name}'))));
-  final experiments = _listNonEmpty(parsedArgs, 'experiments');
+      .map((c) => MapEntry(c, parsedArgs.listNonEmpty('ignore_${c.name}'))));
+  final experiments = parsedArgs.listNonEmpty('experiments');
   final coverageWeb = parsedArgs.flag('coverage_web');
   final healthYamlName = parsedArgs.option('health_yaml_name');
   final healthYamlNames = healthYamlName != null && healthYamlName.isNotEmpty
       ? {healthYamlName}
       : {'health.yaml', 'health.yml'};
+
+  final license = nullIfEmpty(parsedArgs.option('license'));
+  final licenseTestString =
+      nullIfEmpty(parsedArgs.option('license_test_string'));
+
   if (warnOn.toSet().intersection(failOn.toSet()).isNotEmpty) {
     throw ArgumentError('The checks for which warnings are displayed and the '
         'checks which lead to failure must be disjoint.');
@@ -102,10 +117,17 @@ void main(List<String> arguments) async {
     await Health(Directory.current, check, warnOn, failOn, coverageWeb,
             ignorePackages, ignoredFor, experiments, githubApi, flutterPackages,
             healthYamlNames: healthYamlNames,
-            comment: isLocal ? parsedArgs.option('comment') : null)
+            comment: isLocal ? parsedArgs.option('comment') : null,
+    license: license,
+    licenseTestString: licenseTestString,)
         .healthCheck();
   }
 }
 
-List<String> _listNonEmpty(ArgResults parsedArgs, String key) =>
-    (parsedArgs[key] as List<String>).where((e) => e.isNotEmpty).toList();
+String? nullIfEmpty(String? value) =>
+    value != null && value.isNotEmpty ? value : null;
+
+extension on ArgResults {
+  List<String> listNonEmpty(String key) =>
+      (this[key] as List<String>).where((e) => e.isNotEmpty).toList();
+}